Crate debug_filter

Debug Filters

This crate provides a more sophisticated alternative to the common debug level pattern.

The core item are filters. A filter is a string which is divided into components. A filter is similar to a file path. (1)

When you turn allow the filter a, this allows all filters of the form a, a.b, a.b.c, ...

You can also allow a filter like a.b.c, which allows a.b.c, a.b.c.d, ...

Negative filters

There are also negative filters, which disallow the specified filter and all its children. (2)

For example !a.b disallows a.b, a.b.c, ...

Universal filter

There exists a universal filter that allows everything. It is generic, like (1) and (2), and the default is "all"

Multiple filters

You can specify multiple filters at the same time. When two filters apply, the more specific one is chosen.

Example: specifying a.b, !a.b.c and a.b.c.d will allow a.b, a.b.c.d, a.b.e (only if e != c), a.b.e.f, ... and disallow a.b.c, a.b.c.g (only if g != d), a.b.c.g.h, ...

Specifying multiple filters in one string (3) requires another separator, which too is generic and whose default is ':'. The filters from the example would then be a.b:!a.b.c:a.b.c.d

Every filter not listed will be assumed to be disallowed. Include the universal filter in the list to change that.

Footnotes

This crate is generic over the separator for components. The default is '.' so it is used in the examples here.

This crate is generic over the prefix for negating components. The default is '!' so it is used in the examples here.

This is done by the parse method

Macros

const_types
This macro allows quick declaration of a type for ConstValue

declare_filter_parser
A macro to quickly declare a FilterParser

Structs

Filters
A type that stores the filters that are currently active

OwningFilters
Like Filters, but owns the data its keys are from

Enums

DefaultFilterParser

StrAll
A const type for the string "all"

U002E
A const type for the character U+002E Full Stop '.'

U003A
A const type for the character U+003A Colon ':'

U0021
A const type for the character U+0021 Exclamation Mark '!'

Traits

ConstValue
This is a workaround for const generics. This trait assigns a type some constant value.

FilterParser
A trait that includes all the parameters you can customize the parsing behaviour of Filters with at compile time

Type Definitions

Colon

ExclamationMark

FullStop

Point